+++ HISTORY.rst
@@ -8,6 +8,7 @@ Version: 1.4.1 - (UNRELEASED)
 - #304: remove deprecated items from 1.0.0 which were left in place for
   backward compatibility
 - #324: FTPHandler.started attribute, to figure out when client connected.
+- #327: add username and password command line options
 
 **Bug fixes**